src/backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
psycopg[binary]>=3.1
django-cors-headers>=4.3
python-dotenv>=1.0
dj-database-url>=2.1
requests>=2.31
celery>=5.3
redis>=5.0
djangorestframework-simplejwt>=5.3
numpy>=1.26
nibabel>=5.2
